requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
slowapi==0.1.9

# Streamlit frontend
streamlit==1.31.0
streamlit-chat==0.1.1

# OpenAI and embeddings
openai>=1.12.0
tiktoken>=0.8.0

# Vector database
chromadb==0.4.22

# Data processing
pandas==2.2.0
numpy==1.26.3

# Time series forecasting and ML
prophet==1.1.5
statsmodels==0.14.1
scikit-learn==1.4.0

# External APIs
yfinance==0.2.35
requests==2.31.0
fredapi==0.5.1
tavily-python>=0.7.0

# Visualization
plotly==5.18.0
matplotlib>=3.9.0
seaborn==0.13.1

# Security and validation
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]==3.3.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0
